Extract REST poll logic into rest_poll.py
Mirrors metrics_poll.py'''s run(logger) interface so poll.py becomes
a thin, symmetric dispatcher between the two modes instead of inlining
one branch'''s logic directly.
