+++ bin/poll.py
@@ -13,156 +13,17 @@ import os
 
 # Which poller implementation to run: 'rest' (default) polls the Buildkite
 # builds REST API and requires a user API token; 'metrics' polls the agent
-# Metrics API and requires only a cluster agent token (see metrics_poll.py).
-# Deferring the mode-specific imports below (rather than importing both
-# unconditionally) matters: importing `buildkite` eagerly requires a user API
-# token / `.buildkite_token` file, which orgs running metrics-only mode may
-# not have at all.
+# Metrics API and requires only a cluster agent token. Each mode lives in its
+# own module (rest_poll.py / metrics_poll.py), imported lazily here so that
+# rest_poll's eager user-token requirement is never hit when running in
+# metrics-only mode.
 BUILDKITE_POLL_MODE = os.environ.get('BUILDKITE_POLL_MODE', 'rest')
 
-if BUILDKITE_POLL_MODE == 'metrics':
-    import metrics_poll
-    try:
-        metrics_poll.run(logger)
-    except Exception:
-        logger.error("Caught exception during metrics poll", exc_info=True)
-
-else:
-    import re
-    from datetime import date
-    from os.path import join as joinpath
-
-    from buildkite import all_started_builds, all_canceled_builds, build_url
-    from buildkite import get_buildkite_job_tags, BUILDKITE_PATH, BUILDKITE_QUEUE
-    import job_schedulers
-
-    # Time window to query buildkite jobs
-    NHOURS = 96
-
-    # Max concurrent Slurm jobs (pending + running) per buildkite pipeline slug.
-    # All jobs share one service account, so a single pipeline flooding the queue
-    # with high-priority (often un-runnable) jobs can consume the entire per-user
-    # backfill window (bf_max_job_user) and starve every other pipeline. Capping
-    # per-pipeline occupancy keeps the shared window populated with runnable work.
-    # A job over the cap is left 'scheduled' in buildkite and reconsidered next poll.
-    # Per-pipeline overrides go in PIPELINE_LIMITS; everything else uses the default.
-    DEFAULT_PIPELINE_LIMIT = float("inf")  # no cap unless the pipeline is listed below
-    PIPELINE_LIMITS = {
-        "oceananigans-distributed": 3,
-    }
-
-    def pipeline_slug_from_url(url):
-        """Extract the buildkite pipeline slug from a job/build web_url, e.g.
-        'https://buildkite.com/clima/climacoupler-ci/builds/9247#...' -> 'climacoupler-ci'."""
-        match = re.search(r'buildkite\.com/[^/]+/([^/?#]+)', url or '')
-        return match.group(1) if match else None
-
-    try:
-        scheduler = job_schedulers.get_job_scheduler()
-
-        current_jobs = scheduler.current_jobs(logger)
-        logger.info(f"Current jobs (submitted or started): {len(current_jobs)}")
-        logger.debug(f"Current jobs: {current_jobs}")
-
-        # Count active (submitted + running) Slurm jobs per pipeline, so we can
-        # enforce per-pipeline concurrency caps below. Updated in-loop as we submit.
-        pipeline_counts = {}
-        for url, slurm_ids in current_jobs.items():
-            slug = pipeline_slug_from_url(url)
-            if slug:
-                pipeline_counts[slug] = pipeline_counts.get(slug, 0) + len(slurm_ids)
-
-        # poll the buildkite API to check if there are any scheduled/running builds
-        builds = all_started_builds(NHOURS)
-
-        # Accumulate jobs to be canceled in one batch
-        jobs_to_cancel = []
-        # loop over all scheduled and running builds for all pipelines in the buildkite org
-        for build in builds:
-
-            pipeline = build['pipeline']
-            pipeline_name = pipeline['name']
-
-            # for all jobs in this build
-            for job in build['jobs']:
-
-                # jobid, jobtype are attributes in every job object
-                jobid, jobtype = job['id'], job['type']
-
-                # don't schedule non-script jobs on slurm
-                if jobtype != 'script':
-                    continue
-
-                # this job is a script job, check if it has a scheduled state
-                # and not submitted as slurm job
-                # valid states: running, scheduled, passed, failed, blocked,
-                #               canceled, canceling, skipped, not_run, finished
-                # https://buildkite.com/docs/pipelines/defining-steps#build-states
-                jobstate = job['state']
-                buildkite_url = job['web_url']
-
-                # Cancel jobs marked by buildkite as 'canceled'
-                if jobstate == 'canceled':
-                    if buildkite_url in current_jobs:
-                        logger.debug(f"Cancel job: {buildkite_url}")
-                        jobs_to_cancel.append(current_jobs[buildkite_url])
-                    continue
-
-                # jobstate is not pending, or a scheduled job (but not running yet)
-                # is already submitted to slurm
-                if jobstate != 'scheduled' or buildkite_url in current_jobs:
-                    continue
-
-                # Directory containing slurm logs for given build
-                log_dir = joinpath(
-                    BUILDKITE_PATH,
-                    'logs',
-                    f'{date.today()}',
-                    f"build_{build['id']}",
-                )
-
-                job_tags = get_buildkite_job_tags(job)
-                queue = job_tags.get('queue', None)
-
-                # Create the directory prefix if it does not exist
-                if not os.path.isdir(log_dir):
-                    build_link = build_url(pipeline_name, build['number'])
-                    logger.info(f"New build on `{queue}`: {pipeline_name} - {build_link}")
-                    os.mkdir(log_dir)
-
-                # Only log jobs on current queue unless debugging or missing queue
-                if queue is None:
-                    logger.error(f"New job missing queue. Pipeline: {pipeline_name}, {buildkite_url}")
-                    continue
-                elif queue == BUILDKITE_QUEUE:
-                    # Enforce per-pipeline concurrency cap. A deferred job stays
-                    # 'scheduled' in buildkite and is reconsidered on the next poll.
-                    slug = pipeline_slug_from_url(buildkite_url)
-                    limit = PIPELINE_LIMITS.get(slug, DEFAULT_PIPELINE_LIMIT)
-                    if limit is not None and pipeline_counts.get(slug, 0) >= limit:
-                        logger.info(
-                            f"Deferring job, pipeline '{slug}' at cap {limit}: {buildkite_url}"
-                        )
-                        continue
-                    logger.info(f"New job: {pipeline_name}, {buildkite_url}")
-                    scheduler.submit_job(logger, log_dir, job)
-                    # Count this submission so the cap holds within a single poll pass
-                    pipeline_counts[slug] = pipeline_counts.get(slug, 0) + 1
-
-        # Cancel jobs in canceled builds
-        canceled_builds = all_canceled_builds()
-
-        for build in canceled_builds:
-            for job in build['jobs']:
-                if job['type'] == 'script':
-                    buildkite_url = job['web_url']
-                    if buildkite_url in current_jobs:
-                        jobs_to_cancel.append(current_jobs[buildkite_url])
-
-        # Cancel individually marked hpc jobs in one call
-        if jobs_to_cancel:
-            logger.debug(f"Jobs to cancel: {jobs_to_cancel}")
-            scheduler.cancel_jobs(logger, jobs_to_cancel)
-
-    except Exception:
-        logger.error("Caught exception during poll",  exc_info=True)
+try:
+    if BUILDKITE_POLL_MODE == 'metrics':
+        import metrics_poll as poll_impl
+    else:
+        import rest_poll as poll_impl
+    poll_impl.run(logger)
+except Exception:
+    logger.error(f"Caught exception during {BUILDKITE_POLL_MODE} poll", exc_info=True)
